[Artworks] Refactor and add artwork count to section title.
