+++ lib/containers/artist.js
@@ -11,8 +11,6 @@ import Shows from '../components/artist/shows';
 import Articles from '../components/artist/articles';
 import RelatedArtists from '../components/artist/related_artists'
 import Artworks from '../components/artworks';
-import Separator from '../components/separator'
-import SerifText from '../components/text/serif';
 
 import TabView from '../components/tab_view';
 import type TabSelectionEvent from '../components/tab_view';
@@ -40,13 +38,7 @@ class Artist extends React.Component {
         </View>
       );
       case 1: return (
-        <View>
-          <SerifText style={{ marginBottom: 20, fontSize: 20 }}>Works for Sale</SerifText>
-          <Artworks artworks={this.props.artist.for_sale_artworks} paginate={false} />
-          <Separator/>
-          <SerifText style={{ marginBottom: 20, fontSize: 20 }}>Other Works</SerifText>
-          <Artworks artworks={this.props.artist.not_for_sale_artworks} paginate={true} />
-        </View>
+        <Artworks artist={this.props.artist} />
       );
       case 2: return (
         <Shows artist={this.props.artist} />
@@ -81,12 +73,7 @@ export default Relay.createContainer(Artist, {
         articles {
           ${Articles.getFragment('articles')}
         }
-        for_sale_artworks: artworks(sort: published_at_asc, filter: IS_FOR_SALE) {
-          ${Artworks.getFragment('artworks')}
-        }
-        not_for_sale_artworks: artworks(sort: published_at_asc, filter: IS_NOT_FOR_SALE, size: 10) {
-          ${Artworks.getFragment('artworks')}
-        }
+        ${Artworks.getFragment('artist')}
       }
     `,
   }
